Allow more env var configuration
This service can be useful to organize jobs, not only delta-related. A
concrete example I'm having now is: organize initial and healing sync
jobs for a sync between a db and a sharepoint list. Being able to update
more variables allows to better reflect the purpose of the generated
jobs.